Revert "feat: add a topics route handler"
This reverts commit c9d6cc406e066b11b64e3ff2827ed8ce83e6e8f9.
